+++ src/collective/shibboleth/tests/test_example.py
@@ -1,6 +1,6 @@
 import unittest
 
-from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.utils import get_installer
 
 from collective.shibboleth.testing import \
     COLLECTIVE_SHIBBOLETH_INTEGRATION_TESTING
@@ -13,13 +13,12 @@ class TestExample(unittest.TestCase):
     def setUp(self):
         self.app = self.layer['app']
         self.portal = self.layer['portal']
-        self.qi_tool = getToolByName(self.portal, 'portal_quickinstaller')
+        self.installer = get_installer(self.portal, self.layer['request'])
 
     def test_product_is_installed(self):
         """ Validate that our products GS profile has been run and the product
             installed
         """
         pid = 'collective.shibboleth'
-        installed = [p['id'] for p in self.qi_tool.listInstalledProducts()]
-        self.assertTrue(pid in installed,
+        self.assertTrue(self.installer.is_product_installed(pid),
                         'package appears not to have been installed')