+++ src/collective/shibboleth/tests/test_shibbolethloginportlet.py
@@ -1,21 +1,37 @@
-from zope.component import getUtility, getMultiAdapter
-
+from plone import api
+from plone.app.portlets.storage import PortletAssignmentMapping
+from plone.app.testing import TEST_USER_ID
+from plone.app.testing import setRoles
 from plone.portlets.interfaces import IPortletType
 from plone.portlets.interfaces import IPortletManager
 from plone.portlets.interfaces import IPortletAssignment
 from plone.portlets.interfaces import IPortletDataProvider
 from plone.portlets.interfaces import IPortletRenderer
+from zope.component import getUtility, getMultiAdapter
 
-from plone.app.portlets.storage import PortletAssignmentMapping
 
 from collective.shibboleth.portlets import shibbolethloginportlet
-from collective.shibboleth.tests.base_shibbolethloginportlet import TestCase
+# from collective.shibboleth.tests.base_shibbolethloginportlet import TestCase
+
+import unittest
+
+from collective.shibboleth.testing import \
+    COLLECTIVE_SHIBBOLETH_INTEGRATION_TESTING, \
+    COLLECTIVE_SHIBBOLETH_FUNCTIONAL_TESTING
 
 
-class TestPortlet(TestCase):
+class TestPortlet(unittest.TestCase):
 
-    def afterSetUp(self):
-        self.setRoles(('Manager',))
+    layer = COLLECTIVE_SHIBBOLETH_INTEGRATION_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.folder = api.content.create(
+            container=self.portal,
+            type="Folder",
+            title="test"
+        )
 
     def test_portlet_type_registered(self):
         portlet = getUtility(IPortletType, name='collective.shibboleth.portlets.ShibbolethLoginPortlet')
@@ -24,8 +40,8 @@ class TestPortlet(TestCase):
     def test_interfaces(self):
         # TODO: Pass any keywoard arguments to the Assignment constructor
         portlet = shibbolethloginportlet.Assignment()
-        self.failUnless(IPortletAssignment.providedBy(portlet))
-        self.failUnless(IPortletDataProvider.providedBy(portlet.data))
+        self.assertTrue(IPortletAssignment.providedBy(portlet))
+        self.assertTrue(IPortletDataProvider.providedBy(portlet.data))
 
     def test_invoke_add_view(self):
         portlet = getUtility(IPortletType, name='collective.shibboleth.portlets.ShibbolethLoginPortlet')
@@ -38,7 +54,7 @@ class TestPortlet(TestCase):
         addview.createAndAdd(data={})
 
         self.assertEquals(len(mapping), 1)
-        self.failUnless(isinstance(mapping.values()[0], shibbolethloginportlet.Assignment))
+        self.assertTrue(isinstance(mapping.values()[0], shibbolethloginportlet.Assignment))
 
     # NOTE: This test can be removed if the portlet has no edit form
     def test_invoke_edit_view(self):
@@ -47,7 +63,7 @@ class TestPortlet(TestCase):
 
         mapping['foo'] = shibbolethloginportlet.Assignment()
         editview = getMultiAdapter((mapping['foo'], request), name='edit')
-        self.failUnless(isinstance(editview, shibbolethloginportlet.EditForm))
+        self.assertTrue(isinstance(editview, shibbolethloginportlet.EditForm))
 
     def test_obtain_renderer(self):
         context = self.folder
@@ -59,13 +75,21 @@ class TestPortlet(TestCase):
         assignment = shibbolethloginportlet.Assignment()
 
         renderer = getMultiAdapter((context, request, view, manager, assignment), IPortletRenderer)
-        self.failUnless(isinstance(renderer, shibbolethloginportlet.Renderer))
+        self.assertTrue(isinstance(renderer, shibbolethloginportlet.Renderer))
 
 
-class TestRenderer(TestCase):
+class TestRenderer(unittest.TestCase):
 
-    def afterSetUp(self):
-        self.setRoles(('Manager',))
+    layer = COLLECTIVE_SHIBBOLETH_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.folder = api.content.create(
+            container=self.portal,
+            type="Folder",
+            title="test"
+        )
 
     def renderer(self, context=None, request=None, view=None, manager=None, assignment=None):
         context = context or self.folder
@@ -79,16 +103,18 @@ class TestRenderer(TestCase):
 
     def test_render(self):
         # TODO: Pass any keywoard arguments to the Assignment constructor
-        r = self.renderer(context=self.portal, assignment=shibbolethloginportlet.Assignment())
-        r = r.__of__(self.folder)
-        r.update()
-        output = r.render()
+        renderer = self.renderer(context=self.portal, assignment=shibbolethloginportlet.Assignment())
+        renderer.update()
+        # make sure to show the portlet
+        renderer.show = True 
+        output = renderer.render()
         # TODO: Test output
+        self.assertIn("Institutional Login", output)
 
 
-def test_suite():
-    from unittest import TestSuite, makeSuite
-    suite = TestSuite()
-    suite.addTest(makeSuite(TestPortlet))
-    suite.addTest(makeSuite(TestRenderer))
-    return suite
+# def test_suite():
+#     from unittest import TestSuite, makeSuite
+#     suite = TestSuite()
+#     suite.addTest(makeSuite(TestPortlet))
+#     suite.addTest(makeSuite(TestRenderer))
+#     return suite